pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "Smartscope"
authors = [
    {name = "Jonathan Bouvette", email = "bouvette@princeton.edu"},
]
description = "Smartscope module for automatic CryoEM grid screening"
requires-python = ">=3.9,<3.10"
keywords = ["smartscope", "cryoem", "screening","serialem"]
license = {file = "LICENSE"}
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
]
dynamic = ["version", "readme"]
dependencies = [
    "boto3==1.26.157",
    "celery[redis]",
    "channels_redis==4.1.0",
    "channels[daphne]",
    "ddt==1.6.0",
    "django==4.2.2",
    "django-environ==0.10.0",
    "django-filter==23.2",
    "django-storages==1.13.2",
    "djangorestframework==3.14.0",
    "djangorestframework-yaml==2.0.0",
    "drawSvg[all]",
    "environ==1.0",
    "gunicorn",
    "imutils==0.5.4",
    "mrcfile==1.4.3",
    "matplotlib==3.7.1",
    "mysqlclient",
    "numpy==1.25.0",
    "opencv-python==4.7.0.72",
    "pandas==2.0.2",
    "pillow==9.0.1",
    "plotly==5.15.0",
    "pycv2==0.0.1",
    "pydantic",
    "psutil==5.9.5",
    "pymysql==1.0.3",
    "pyyaml==6.0",
    "seaborn",
    "scipy==1.10.1",
    "scikit-image==0.21.0",
    "scikit-learn",
    "torch==1.8.1",
    "torchvision==0.9.1",
    "torchaudio==0.8.1",
    "uvicorn",
    "websockets==11.0.3",
]


[project.optional-dependencies]
dev = ["black", "flake8"]

[project.urls]
Homepage = "https://smartscope.org"
Documentation = "https://docs.smartscope.org/"
Repository = "https://github.com/NIEHS/SmartScope"
# [project.scripts]
# my-script = "my_package.module:function"

[tool.setuptools.dynamic]
version = {file = ["VERSION"]}
readme = {file = ["README.md"]}

[tool.setuptools]
packages = ["Smartscope"]
script-files = ["Smartscope/bin/smartscope.py","Smartscope/bin/manage.py","Smartscope/bin/smartscope.sh"]